Screens for CimiTargets #36

+++ src/biz.dfch.CS.Appclusive.UI/Controllers/CimiTargetsController.cs
@@ -35,7 +35,7 @@ namespace biz.dfch.CS.Appclusive.UI.Controllers
         }
 
         // GET: CimiTargets/Details/5
-        public ActionResult Details(int id)
+        public ActionResult Details(long id)
         {
             try
             {
